added the boot receiver class for setting up things we do routinely anytime app starts up. boot receiver class is disabled and only enabled when user signs up or login. it is also disabled when user signs out.
fixed a bug in the contact manager class that makes it return no contact when no filter is passed to it.

also fixed a bug in ChatActivity where unbind is called twice.

**************this should have been added to a previous commit but forgot*********************

also added a functionality to MessageDispatcher where all jobs are counted and the passed monitor notified when all jobs completed.

Message Dispatcher is now package private