online, typing,offline etc added with full test. typing announcement handling when client is not in the chat room. Status Manager Cleanup added MainThreadBaseEventListener
PairAppClient now uses a Set to hold activities instead of a stack. this makes it possible for us to detect online, offline efficiently by notifying other clients we are online in onResume and in onPause we say we are offline. previously this was not easily done with a stack.

as usual tests were updated accordingly